dashboard/requirements.txt
streamlit
plotly
pandas
supabase==1.2.0
python-dotenv